DDPB-797 Moved notes association from report to client
